mcp_server/requirements.txt
mcp>=1.0.0
Flask==3.1.1
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
sqlalchemy-utils>=0.41.1
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0